move: dialog and snackbar to feedback directory
